move fixture into conftest
