+++ tests/data/test_grids.py
@@ -1,30 +1,4 @@
-from pathlib import Path
-import pytest
-
-import numpy as np
 from toshi_hazard_store.model import ProbabilityEnum
-
-from nzshm_hazlab.data import HazardGrids
-
-
-class DummyGridLoader:
-    """A dummy grid loader class."""
-
-    def __init__(self, directory: Path | str):
-        self._directory = Path(directory)
-
-    def get_grid(self, hazard_model_id, imt, grid_name, vs30, poe, agg: str):
-        """A method for the dummy loader."""
-        filename = f"hazard_grid-{hazard_model_id}-{imt}-{grid_name}-{vs30}-{poe.name}-{agg}.npy"
-        filepath = self._directory / filename
-        return np.load(filepath)
-
-# scope is function because we attach a mocker.spy to the object which only seems to work for function scope
-@pytest.fixture(scope='function')
-def hazard_grids() -> HazardGrids:
-    loader = DummyGridLoader(Path(__file__).parent.parent / "fixtures/data/grids/")
-    return HazardGrids(loader=loader)
-
 
 
 def test_get_grid(hazard_grids):
@@ -34,7 +8,6 @@ def test_get_grid(hazard_grids):
     agg = "mean"
     imt = "SA(1.0)"
     vs30 = 750
-
 
     imtls = hazard_grids.get_grid(hazard_model_id, imt, grid_name, vs30, poe, agg)
     assert imtls.shape == (3741,)
